pyproject.toml
[project]
name = "triadllm"
version = "0.1.0"
description = "Terminal UI for proposal-validation-consolidation LLM workflows across OpenAI, Mistral, and compatible backends."
readme = "README.md"
authors = [
    { name = "David R. Lopez B", email = "ibitato@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.13,<3.14"
dependencies = [
    "mistralai>=1.12.4",
    "openai>=2.26.0",
    "platformdirs>=4.9.4",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "textual>=8.0.2",
]

[project.scripts]
triad = "triadllm:main"
triadllm = "triadllm:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
